+++ lib/src/const/constants.dart
@@ -1,0 +1,5 @@
+// Copyright (c) 2019 Ben Hills and the project contributors. Use of this source
+// code is governed by a MIT license that can be found in the LICENSE file.
+
+const podcastSearchAgent =
+    'podcast_search/0.7.8 https://github.com/amugofjava/anytime_podcast_player';